+++ src/cli/submit/stdin/Stdin.Handler.ts
@@ -14,22 +14,12 @@ import { JobUtils } from "../../../api/JobUtils";
 import { FTPBaseHandler } from "../../../FTPBase.Handler";
 import { IFTPHandlerParams } from "../../../IFTPHandlerParams";
 import { CoreUtils } from "../../../api/CoreUtils";
+import { SubmitJobHandler } from "../SubmitJobHandler";
 
-export default class SubmitJobFromLocalFileHandler extends FTPBaseHandler {
+export default class SubmitJobFromLocalFileHandler extends SubmitJobHandler {
     public async processFTP(params: IFTPHandlerParams): Promise<void> {
-        let jobDetails: any;
-        let jobid: string;
         this.log.debug("Submitting a job from stdin %s");
         const stdinContent = await CoreUtils.readStdin();
-        jobid = await params.connection.submitJCL(stdinContent.toString());
-        jobDetails = await JobUtils.findJobByID(jobid, params.connection);
-        this.log.info("Submitted job successfully, jobname(jobid): %s(%s)", jobDetails.jobname, jobDetails.jobid);
-        params.response.data.setObj(jobDetails);
-        params.response.format.output({
-            output: jobDetails,
-            format: "object",
-            fields: ["jobid", "jobname", "owner", "status"]
-        });
+        return this.submitJCL(stdinContent.toString(), params);
     }
 }
-
